Implemented required use cases; improvements still to be made

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -3,23 +3,23 @@ package com.techelevator.tenmo.model;
 import java.math.BigDecimal;
 
 public class Account {
-    private int accountId;
-    private int userId;
+    private long accountId;
+    private long userId;
     private BigDecimal balance;
 
-    public int getAccountId() {
+    public long getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(int accountId) {
+    public void setAccountId(long accountId) {
         this.accountId = accountId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
