+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,11 +4,11 @@ import java.math.BigDecimal;
 
 public class Account {
 
-    private int accountId;
-    private int userId;
+    private long accountId;
+    private long userId;
     private BigDecimal balance;
 
-    public Account(int accountId, int userId, BigDecimal balance){
+    public Account(long accountId, long userId, BigDecimal balance){
         this.accountId = accountId;
         this.userId = userId;
         this.balance = balance;
@@ -16,19 +16,19 @@ public class Account {
 
     public Account() { }
 
-    public int getAccountId() {
+    public long getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(int accountId) {
+    public void setAccountId(long accountId) {
         this.accountId = accountId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
